Add files via upload
The python program is used to simulate the summary statistics expected according to the parameters of the model (birth and death rate, competition kernel, mutation probabilities and kernel). The python program compressed in Programme_2traits.rar can be launched through the file simulationsLauncher.py where one can define the number of simulations required and whether parameters should be drawn from distributions or should be fixed values (details of the distributions and values can be set in the file getParameters.py). The demographic equations are described in lotkaVolterra.py.